Make a more robust build command - nukebuild.sh
The build typically fails if no caching service exists, or for the first
run on eash stage with apt-cacher-ng. Also there is a new wrinkle - if
you use SKIP to start with the currently failing stage, "patch" dirs
fail. Make a new command that nukes the work directory and starts over
if a failure is encountered.